Do not use Java 11 APIs

+++ src/main/kotlin/org/danilopianini/upgradle/modules/RefreshVersions.kt
@@ -11,7 +11,6 @@ import java.io.File
 import java.io.InputStream
 import java.nio.charset.StandardCharsets
 import java.util.concurrent.TimeUnit
-import java.util.concurrent.TimeoutException
 
 class RefreshVersions(options: Map<String, Any>) : GradleRootModule() {
 
@@ -106,18 +105,16 @@ class RefreshVersions(options: Map<String, Any>) : GradleRootModule() {
         val process = ProcessBuilder(listOf(gradleCommand, "--console=plain", taskName))
             .directory(projectRoot)
             .start()
-        return try {
-            when (process.onExit().get(timeoutInMinutes, TimeUnit.MINUTES).exitValue()) {
+        return if (process.waitFor(timeoutInMinutes, TimeUnit.MINUTES)) {
+            when (process.exitValue()) {
                 0 -> ProcessOutcome.Ok
                 else -> ProcessOutcome.Error(process)
             }
-        } catch (timeOut: TimeoutException) {
-            logger.error("One process has timed out", timeOut)
+        } else {
+            logger.error("One process has timed out")
             process.destroy()
-            try {
-                process.onExit().get(1L, TimeUnit.SECONDS)
-            } catch (unkillableTimeOut: TimeoutException) {
-                logger.error("One process refused to terminate after timeout, destroying forcibly", unkillableTimeOut)
+            if (!process.waitFor(1L, TimeUnit.SECONDS)) {
+                logger.error("One process refused to terminate after timeout, destroying forcibly")
                 process.destroyForcibly()
             }
             ProcessOutcome.TimeOut(process)
